+++ doc/change-log.md
@@ -13,10 +13,11 @@
 - 增加 `DERBY` 方言
 - `SqlDialect` 增加 `isPostgreSQL`, `isDerby`, `isSqlite` 方法
 - 引入 `com.lucendar.common.db.jdbc.DbHelper.timestampStmtSetter`
-- 引入 `com.lucendar.common.db.jdbc.ResultSetAccessor.date()`, `timestamp()` 方法
+- 引入 `com.lucendar.common.db.jdbc.ResultSetAccessor.date()`, `timestamp()`, `beijingConvTimestampStr()` 方法
 - 引入 `com.lucendar.common.db.jdbc.StatementBinder.setDate` 方法
 - 引入 `com.lucendar.common.db.jdbc.DbHelper.dateStmtSetter` 方法
-- 引入 `com.lucendar.common.db.jdbc.DbSupport.tableExists(String, String)` 方法 
+- 引入 `com.lucendar.common.db.jdbc.DbSupport.tableExists(String, String)` 方法
+- `SimpleSelect` 增加 `sqlDialect` 字段
 
 ## 变更
 - 不兼容变更